refactor(validate): split description_yml into focused submodules
The 957-line flat file is replaced with a directory module:

- model.rs     — DescriptionYml struct (~40 lines)
- parser.rs    — parse_description_yml + private parse_kv /
                 strip_inline_comment helpers (~230 lines)
- validator.rs — validate_description_yml_str + validate_rust_extension
                 (~100 lines)
- tests.rs     — all 20 unit tests extracted from mod tests (~250 lines)
- mod.rs       — module doc + public re-exports (~80 lines)

Public API is unchanged; all 303 unit, 40 integration, and 90 doc tests pass.
